# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Irish Payroll Management & Compliance System",
  "description": "Portfolio prototype. Statutory values are versioned rules with official sources; unverified rules are blocked from calculation. Not certified payroll software.",
  "version": "1.0.0"
 },
 "paths": {
  "/auth/token": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_auth_token_post",
    "requestBody": {
     "content": {
      "application/x-www-form-urlencoded": {
       "schema": {
        "$ref": "#/components/schemas/Body_login_auth_token_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Token"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/companies": {
   "post": {
    "tags": [
     "companies"
    ],
    "summary": "Create Company",
    "operationId": "create_company_companies_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompanyIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CompanyOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/companies/{company_id}": {
   "get": {
    "tags": [
     "companies"
    ],
    "summary": "Get Company",
    "operationId": "get_company_companies__company_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "company_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Company Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CompanyOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/employees": {
   "post": {
    "tags": [
     "employees"
    ],
    "summary": "Add Employee",
    "operationId": "add_employee_employees_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EmployeeIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EmployeeOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "employees"
    ],
    "summary": "List Employees",
    "operationId": "list_employees_employees_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "company_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company Id"
      }
     },
     {
      "name": "search",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Search"
      }
     },
     {
      "name": "department",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Department"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 500,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/EmployeeOut"
         },
         "title": "Response List Employees Employees Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/employees/{employee_id}": {
   "get": {
    "tags": [
     "employees"
    ],
    "summary": "Get Employee",
    "operationId": "get_employee_employees__employee_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "employee_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Employee Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "employees"
    ],
    "summary": "Patch Employee",
    "operationId": "patch_employee_employees__employee_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "employee_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Employee Id"
      }
     },
     {
      "name": "reason",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Reason"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true,
        "title": "Changes"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EmployeeOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/rpn/import": {
   "post": {
    "tags": [
     "rpn"
    ],
    "summary": "Rpn Import",
    "operationId": "rpn_import_rpn_import_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "company_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Company Id"
      }
     },
     {
      "name": "tax_year",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Tax Year"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_rpn_import_rpn_import_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Body_login_auth_token_post": {
    "properties": {
     "grant_type": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^password$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grant Type"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "format": "password",
      "title": "Password"
     },
     "scope": {
      "type": "string",
      "title": "Scope",
      "default": ""
     },
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id"
     },
     "client_secret": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "format": "password",
      "title": "Client Secret"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "Body_login_auth_token_post"
   },
   "Body_rpn_import_rpn_import_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_rpn_import_rpn_import_post"
   },
   "CompanyIn": {
    "properties": {
     "legal_name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 2,
      "title": "Legal Name"
     },
     "trading_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trading Name"
     },
     "tax_registration_number": {
      "type": "string",
      "maxLength": 20,
      "minLength": 7,
      "title": "Tax Registration Number"
     },
     "employer_registration_details": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employer Registration Details"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "contact_email": {
      "anyOf": [
       {
        "type": "string",
        "format": "email"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Email"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "payroll_frequency": {
      "type": "string",
      "enum": [
       "WEEKLY",
       "FORTNIGHTLY",
       "MONTHLY"
      ],
      "title": "Payroll Frequency",
      "default": "MONTHLY"
     },
     "financial_year": {
      "type": "integer",
      "maximum": 2030.0,
      "minimum": 2025.0,
      "title": "Financial Year"
     }
    },
    "type": "object",
    "required": [
     "legal_name",
     "tax_registration_number",
     "financial_year"
    ],
    "title": "CompanyIn"
   },
   "CompanyOut": {
    "properties": {
     "legal_name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 2,
      "title": "Legal Name"
     },
     "trading_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trading Name"
     },
     "tax_registration_number": {
      "type": "string",
      "maxLength": 20,
      "minLength": 7,
      "title": "Tax Registration Number"
     },
     "employer_registration_details": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employer Registration Details"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "contact_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Email"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "payroll_frequency": {
      "type": "string",
      "enum": [
       "WEEKLY",
       "FORTNIGHTLY",
       "MONTHLY"
      ],
      "title": "Payroll Frequency",
      "default": "MONTHLY"
     },
     "financial_year": {
      "type": "integer",
      "maximum": 2030.0,
      "minimum": 2025.0,
      "title": "Financial Year"
     },
     "company_id": {
      "type": "integer",
      "title": "Company Id"
     }
    },
    "type": "object",
    "required": [
     "legal_name",
     "tax_registration_number",
     "financial_year",
     "company_id"
    ],
    "title": "CompanyOut"
   },
...
   "EmployeeIn": {
    "properties": {
     "company_id": {
      "type": "integer",
      "title": "Company Id"
     },
     "employee_number": {
      "type": "string",
      "maxLength": 20,
      "minLength": 1,
      "title": "Employee Number"
     },
     "first_name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "First Name"
     },
     "last_name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Last Name"
     },
     "ppsn": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ppsn"
     },
     "date_of_birth": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date Of Birth"
     },
     "employment_start_date": {
      "type": "string",
      "format": "date",
      "title": "Employment Start Date"
     },
     "employment_end_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employment End Date"
     },
     "department_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Department Id"
     },
     "job_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Title"
     },
     "location": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "contract_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contract Type"
     },
     "salary_type": {
      "type": "string",
      "enum": [
       "SALARY",
       "HOURLY"
      ],
      "title": "Salary Type",
      "default": "SALARY"
     },
     "annual_salary": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Annual Salary"
     },
     "hourly_rate": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hourly Rate"
     },
     "standard_hours": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 168.0,
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Standard Hours"
     },
     "pay_frequency": {
      "type": "string",
      "enum": [
       "WEEKLY",
       "FORTNIGHTLY",
       "MONTHLY"
      ],
      "title": "Pay Frequency",
      "default": "MONTHLY"
     },
     "prsi_class": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2
       },
       {
        "type": "null"
       }
      ],
      "title": "Prsi Class",
      "default": "A"
     },
     "pension_scheme": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "OCCUPATIONAL",
         "AVC",
         "PRSA",
         "RAC"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Pension Scheme"
     },
     "pension_ee_percent": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 1.0,
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pension Ee Percent"
     },
     "pension_er_percent": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 1.0,
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pension Er Percent"
     },
     "employment_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employment Id"
     }
    },
    "type": "object",
    "required": [
     "company_id",
     "employee_number",
     "first_name",
     "last_name",
     "employment_start_date"
    ],
    "title": "EmployeeIn"
   },
   "EmployeeOut": {
    "properties": {
     "employee_id": {
      "type": "integer",
      "title": "Employee Id"
     },
     "employee_number": {
      "type": "string",
      "title": "Employee Number"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "ppsn_masked": {
      "type": "string",
      "title": "Ppsn Masked"
     },
     "department": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Department"
     },
     "job_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Title"
     },
     "pay_frequency": {
      "type": "string",
      "title": "Pay Frequency"
     },
     "salary_type": {
      "type": "string",
      "title": "Salary Type"
     },
     "annual_salary": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Annual Salary"
     },
     "prsi_class": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prsi Class"
     },
     "employment_status": {
      "type": "string",
      "title": "Employment Status"
     },
     "has_rpn": {
      "type": "boolean",
      "title": "Has Rpn"
     }
    },
    "type": "object",
    "required": [
     "employee_id",
     "employee_number",
     "name",
     "ppsn_masked",
     "department",
     "job_title",
     "pay_frequency",
     "salary_type",
     "annual_salary",
     "prsi_class",
     "employment_status",
     "has_rpn"
    ],
    "title": "EmployeeOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Token": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "role": {
      "type": "string",
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "role"
    ],
    "title": "Token"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/auth/token"
     }
    }
   }
  }
 }
}